style: split Salary and Correction page into smaller components
